Move DeallocationSample destructor to .cc file (#6)
DeallocationSample holds a std::unique_ptr to a forward-declared class
that is defined in the corresponding .cc file (similar to pimpl).
Definining the destructor inline in the header is problematic since
std::unique_ptr destructor requires the class to be defined, not just
declared. This appears as a compile failure when compiling with clang
17/19 and --std=c++23.

This change moves the destructor definition into the source file.